refactor: akazeテスト用のエンドポイントの移動と不要ファイルの削除をしたよ
