+++ api/api/main.py
@@ -1,6 +1,5 @@
 from fastapi import FastAPI
 from api.routers import picture2item, event, auth, mobile
-from api.routers.events import match
 from fastapi.middleware.cors import CORSMiddleware
 
 app = FastAPI()
@@ -23,7 +22,6 @@ async def root():
     return {"message": "Hello World"}
 
 app.include_router(picture2item.router)
-app.include_router(match.router)
 app.include_router(event.router)
 app.include_router(auth.router)
 app.include_router(mobile.router)
